requirements.txt
gradio==4.44.0
datasets>=2.19.0
huggingface-hub>=0.20
faiss-cpu==1.7.4
numpy<2
python-dotenv
openai>=1.0.0